Restore missing documentation comments for split cache files
