+++ Assets/Aramaa/OchibiChansConverterTool/Editor/UI/Components/OCTPrefabDropdownCache.FaceMeshMatching.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 286a58c4ef7b2578e29dfa51ab637315
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
