+++ Assets/Aramaa/OchibiChansConverterTool/Editor/UI/Components/OCTPrefabDropdownCache.Models.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7df4d284664a9ce2f003ad9c30795909
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
