+++ Assets/Aramaa/OchibiChansConverterTool/Editor/UI/Components/OCTPrefabDropdownCache.cs
@@ -12,6 +12,18 @@
 // - Project 側の Prefab を読み取るだけで、Prefab 自体の内容は変更しません
 //
 // ============================================================================
+// ファイル分割メモ
+// ============================================================================
+// - OCTPrefabDropdownCache.cs
+//   候補一覧の状態管理・公開 API・候補一覧の構築フロー
+// - OCTPrefabDropdownCache.FaceMeshMatching.cs
+//   顔メッシュの抽出・一致判定・キャッシュ利用の中核ロジック
+// - OCTPrefabDropdownCache.Persistence.cs
+//   Library 配下キャッシュの読み書き
+// - OCTPrefabDropdownCache.Models.cs
+//   判定と保存で共有する内部データモデル
+//
+// ============================================================================
 // チーム開発向けルール
 // ============================================================================
 // - キャッシュは Library 配下に保存し、VCS には含めません（各環境ローカル）
@@ -25,16 +37,13 @@ using System.IO;
 using System.Linq;
 using UnityEditor;
 using UnityEngine;
-#if VRC_SDK_VRCSDK3
-using VRC.SDK3.Avatars.Components;
-#endif
 
 namespace Aramaa.OchibiChansConverterTool.Editor
 {
     /// <summary>
     /// アバターの顔メッシュ情報を基に、候補となるおちびちゃんズ Prefab の一覧を作るキャッシュです。
     /// </summary>
-    internal sealed class OCTPrefabDropdownCache
+    internal sealed partial class OCTPrefabDropdownCache
     {
         private static string L(string key) => OCTLocalization.Get(key);
         private static string F(string key, params object[] args) => OCTLocalization.Format(key, args);
@@ -240,466 +249,6 @@ namespace Aramaa.OchibiChansConverterTool.Editor
                     .IndexOf(pattern, StringComparison.OrdinalIgnoreCase) >= 0);
 
             return match;
-        }
-
-        private static bool PrefabHasMatchingFaceMesh(string prefabPath, FaceMeshSignature targetFaceMeshSignature)
-        {
-            if (string.IsNullOrEmpty(prefabPath)) return false;
-            if (!targetFaceMeshSignature.HasAnyIdentity) return false;
-
-            return TryGetCachedFaceMeshSignature(prefabPath, out var prefabFaceMeshSignature) &&
-                   FaceMeshSignatureMatches(targetFaceMeshSignature, prefabFaceMeshSignature);
-        }
-
-        /// <summary>
-        /// アバターの Viseme 用メッシュから、GUID/LocalId を抽出します。
-        /// </summary>
-        private static bool TryGetFaceMeshSignature(GameObject root, out FaceMeshSignature signature)
-        {
-            signature = default;
-            if (root == null) return false;
-
-#if VRC_SDK_VRCSDK3
-            var descriptor = root.GetComponent<VRC.SDK3.Avatars.Components.VRCAvatarDescriptor>();
-            if (!TryGetVisemeRendererFromDescriptor(descriptor, out var faceRenderer)) return false;
-            if (faceRenderer == null || faceRenderer.sharedMesh == null) return false;
-
-            var mesh = faceRenderer.sharedMesh;
-            if (!TryGetPrefabInfo(root, out var prefabGuid, out var prefabName))
-            {
-                prefabGuid = string.Empty;
-                prefabName = string.Empty;
-            }
-
-            return TryBuildFaceMeshSignature(mesh, prefabGuid, prefabName, out signature);
-#else
-            return false;
-#endif
-        }
-
-        private static bool TryGetFaceMeshSignatureFromPrefabPath(string prefabPath, out FaceMeshSignature signature)
-        {
-            signature = default;
-            if (string.IsNullOrEmpty(prefabPath)) return false;
-
-            var prefab = AssetDatabase.LoadMainAssetAtPath(prefabPath) as GameObject;
-            if (prefab == null) return false;
-
-            if (!TryGetFaceMeshSignature(prefab, out signature)) return false;
-
-            var prefabSource = PrefabUtility.GetCorrespondingObjectFromOriginalSource(prefab) ?? prefab;
-            var sourcePath = AssetDatabase.GetAssetPath(prefabSource);
-            if (string.IsNullOrEmpty(sourcePath))
-            {
-                sourcePath = prefabPath;
-            }
-
-            var prefabGuid = AssetDatabase.AssetPathToGUID(sourcePath);
-            var prefabName = sourcePath;
-            signature = signature.WithPrefabInfo(prefabGuid, prefabName);
-            return true;
-        }
-
-        /// <summary>
-        /// Prefab の依存ハッシュを使って、顔メッシュIDのキャッシュを再利用します。
-        /// </summary>
-        private static bool TryGetCachedFaceMeshSignature(string prefabPath, out FaceMeshSignature signature)
-        {
-            signature = default;
-            if (string.IsNullOrEmpty(prefabPath)) return false;
-
-            EnsureFaceMeshCacheLoaded();
-
-            var hash = AssetDatabase.GetAssetDependencyHash(prefabPath);
-            if (CachedFaceMeshByPrefab.TryGetValue(prefabPath, out var cached) &&
-                cached.DependencyHash == hash)
-            {
-                if (cached.HasFaceMesh)
-                {
-                    signature = cached.FaceMeshSignature;
-                    return true;
-                }
-
-                return false;
-            }
-
-            var hasFaceMesh = TryGetFaceMeshSignatureFromPrefabPath(prefabPath, out var cachedSignature);
-            CachedFaceMeshByPrefab[prefabPath] = new CachedFaceMesh(hash, cachedSignature, hasFaceMesh);
-            MarkFaceMeshCacheDirty();
-            // ここでは即時保存しません。
-            // - OnDisable でまとめて保存される
-            // - 検索中に毎回ディスク/設定を書き換える回数を減らし、Editor の負荷を抑える
-            if (hasFaceMesh)
-            {
-                signature = cachedSignature;
-            }
-
-            return hasFaceMesh;
-        }
-
-        private static bool FaceMeshSignatureMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (MeshIdMatches(a.MeshId, b.MeshId))
-            {
-                return true;
-            }
-
-            if (PrefabGuidMatches(a, b)) return true;
-            if (PrefabNameMatches(a, b)) return true;
-            if (FbxGuidMatches(a, b)) return true;
-            if (FbxNameMatches(a, b)) return true;
-            if (AssetPathMatches(a, b)) return true;
-
-            return false;
-        }
-
-        private static bool MeshIdMatches(MeshId a, MeshId b)
-        {
-            if (string.IsNullOrEmpty(a.Guid) || string.IsNullOrEmpty(b.Guid)) return false;
-            if (!string.Equals(a.Guid, b.Guid, StringComparison.Ordinal)) return false;
-
-            if (a.HasLocalId && b.HasLocalId)
-            {
-                return a.LocalId == b.LocalId;
-            }
-
-            return true;
-        }
-
-        private static bool PrefabGuidMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (string.IsNullOrEmpty(a.PrefabGuid) || string.IsNullOrEmpty(b.PrefabGuid)) return false;
-            return string.Equals(a.PrefabGuid, b.PrefabGuid, StringComparison.Ordinal);
-        }
-
-        private static bool PrefabNameMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (string.IsNullOrEmpty(a.PrefabName) || string.IsNullOrEmpty(b.PrefabName)) return false;
-            return string.Equals(a.PrefabName, b.PrefabName, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static bool FbxGuidMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (string.IsNullOrEmpty(a.FbxGuid) || string.IsNullOrEmpty(b.FbxGuid)) return false;
-            return string.Equals(a.FbxGuid, b.FbxGuid, StringComparison.Ordinal);
-        }
-
-        private static bool FbxNameMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (string.IsNullOrEmpty(a.FbxName) || string.IsNullOrEmpty(b.FbxName)) return false;
-            return string.Equals(a.FbxName, b.FbxName, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static bool AssetPathMatches(FaceMeshSignature a, FaceMeshSignature b)
-        {
-            if (string.IsNullOrEmpty(a.FaceMeshAssetPath) || string.IsNullOrEmpty(b.FaceMeshAssetPath)) return false;
-            return string.Equals(a.FaceMeshAssetPath, b.FaceMeshAssetPath, StringComparison.OrdinalIgnoreCase);
-        }
-
-        private static bool TryGetPrefabInfo(GameObject root, out string prefabGuid, out string prefabName)
-        {
-            prefabGuid = string.Empty;
-            prefabName = string.Empty;
-            if (root == null) return false;
-
-            var instanceRoot = PrefabUtility.GetNearestPrefabInstanceRoot(root);
-            if (instanceRoot == null) return false;
-
-            var prefabAsset = PrefabUtility.GetCorrespondingObjectFromSource(instanceRoot);
-            if (prefabAsset == null) return false;
-
-            var sourceAsset = PrefabUtility.GetCorrespondingObjectFromOriginalSource(prefabAsset) ?? prefabAsset;
-            var prefabPath = AssetDatabase.GetAssetPath(sourceAsset);
-            if (string.IsNullOrEmpty(prefabPath)) return false;
-
-            prefabGuid = AssetDatabase.AssetPathToGUID(prefabPath);
-            prefabName = prefabPath;
-            return !string.IsNullOrEmpty(prefabGuid) || !string.IsNullOrEmpty(prefabName);
-        }
-
-        private static bool TryBuildFaceMeshSignature(
-            Mesh mesh,
-            string prefabGuid,
-            string prefabName,
-            out FaceMeshSignature signature)
-        {
-            signature = default;
-            if (mesh == null) return false;
-
-            var assetPath = AssetDatabase.GetAssetPath(mesh);
-            var fbxGuid = string.IsNullOrEmpty(assetPath) ? string.Empty : AssetDatabase.AssetPathToGUID(assetPath);
-            var fbxName = string.IsNullOrEmpty(assetPath) ? string.Empty : assetPath;
-            var hasMeshId = TryBuildMeshId(mesh, out var meshId);
-
-            if (!hasMeshId &&
-                string.IsNullOrEmpty(prefabGuid) &&
-                string.IsNullOrEmpty(prefabName) &&
-                string.IsNullOrEmpty(fbxGuid) &&
-                string.IsNullOrEmpty(fbxName) &&
-                string.IsNullOrEmpty(assetPath))
-            {
-                return false;
-            }
-
-            signature = new FaceMeshSignature(meshId, prefabGuid, prefabName, fbxGuid, fbxName, assetPath);
-            return true;
-        }
-
-        private static bool TryBuildMeshId(Mesh mesh, out MeshId meshId)
-        {
-            meshId = default;
-            if (mesh == null) return false;
-
-            if (AssetDatabase.TryGetGUIDAndLocalFileIdentifier(mesh, out var guid, out long localId))
-            {
-                meshId = new MeshId(guid, localId, hasLocalId: true);
-                return true;
-            }
-
-            var meshPath = AssetDatabase.GetAssetPath(mesh);
-            if (string.IsNullOrEmpty(meshPath)) return false;
-
-            var fallbackGuid = AssetDatabase.AssetPathToGUID(meshPath);
-            if (string.IsNullOrEmpty(fallbackGuid)) return false;
-
-            meshId = new MeshId(fallbackGuid, 0, hasLocalId: false);
-            return true;
-        }
-
-#if VRC_SDK_VRCSDK3
-        private static bool TryGetVisemeRendererFromDescriptor(
-            VRCAvatarDescriptor descriptor,
-            out SkinnedMeshRenderer renderer)
-        {
-            renderer = null;
-            if (descriptor == null) return false;
-
-            using (var so = new SerializedObject(descriptor))
-            {
-                var prop = so.FindProperty("VisemeSkinnedMesh");
-                if (prop != null)
-                {
-                    renderer = prop.objectReferenceValue as SkinnedMeshRenderer;
-                }
-            }
-
-            if (renderer != null) return true;
-
-            renderer = descriptor.VisemeSkinnedMesh;
-            return renderer != null;
-        }
-#endif
-
-        private static void EnsureFaceMeshCacheLoaded()
-        {
-            if (_faceMeshCacheLoaded) return;
-            _faceMeshCacheLoaded = true;
-            LoadFaceMeshCacheFromLibrary();
-        }
-
-        private static void MarkFaceMeshCacheDirty()
-        {
-            _faceMeshCacheDirty = true;
-        }
-
-        private static void LoadFaceMeshCacheFromLibrary()
-        {
-            try
-            {
-                var cachePath = GetFaceMeshCacheFilePath();
-                if (!File.Exists(cachePath)) return;
-
-                var json = File.ReadAllText(cachePath);
-                if (string.IsNullOrEmpty(json)) return;
-
-                var cacheFile = JsonUtility.FromJson<FaceMeshCacheFile>(json);
-                if (cacheFile == null) return;
-                if (cacheFile.Entries == null) return;
-
-                foreach (var entry in cacheFile.Entries)
-                {
-                    if (entry == null) continue;
-                    if (string.IsNullOrEmpty(entry.PrefabPath)) continue;
-                    if (string.IsNullOrEmpty(entry.DependencyHash)) continue;
-
-                    if (!TryParseHash128(entry.DependencyHash, out var hash)) continue;
-
-                    var meshId = new MeshId(entry.FaceMeshGuid ?? string.Empty, entry.FaceMeshLocalId, entry.HasLocalId);
-                    var signature = new FaceMeshSignature(
-                        meshId,
-                        entry.PrefabGuid ?? string.Empty,
-                        entry.PrefabName ?? string.Empty,
-                        entry.FbxGuid ?? string.Empty,
-                        entry.FbxName ?? string.Empty,
-                        entry.FaceMeshAssetPath ?? string.Empty);
-                    CachedFaceMeshByPrefab[entry.PrefabPath] = new CachedFaceMesh(hash, signature, entry.HasFaceMesh);
-                }
-            }
-            catch (Exception e)
-            {
-                Debug.LogWarning(F("Warning.FaceMeshCacheLoadFailed", e.Message));
-            }
-        }
-
-        private static void SaveFaceMeshCacheToLibrary()
-        {
-            if (!_faceMeshCacheDirty) return;
-
-            try
-            {
-                var cacheFile = new FaceMeshCacheFile();
-                foreach (var pair in CachedFaceMeshByPrefab)
-                {
-                    var cached = pair.Value;
-                    cacheFile.Entries.Add(new FaceMeshCacheEntry
-                    {
-                        PrefabPath = pair.Key,
-                        DependencyHash = cached.DependencyHash.ToString(),
-                        FaceMeshGuid = cached.FaceMeshSignature.MeshId.Guid,
-                        FaceMeshLocalId = cached.FaceMeshSignature.MeshId.LocalId,
-                        HasLocalId = cached.FaceMeshSignature.MeshId.HasLocalId,
-                        PrefabGuid = cached.FaceMeshSignature.PrefabGuid,
-                        PrefabName = cached.FaceMeshSignature.PrefabName,
-                        FbxGuid = cached.FaceMeshSignature.FbxGuid,
-                        FbxName = cached.FaceMeshSignature.FbxName,
-                        FaceMeshAssetPath = cached.FaceMeshSignature.FaceMeshAssetPath,
-                        HasFaceMesh = cached.HasFaceMesh
-                    });
-                }
-
-                var json = JsonUtility.ToJson(cacheFile, true);
-                var cachePath = GetFaceMeshCacheFilePath();
-                Directory.CreateDirectory(Path.GetDirectoryName(cachePath) ?? string.Empty);
-                File.WriteAllText(cachePath, json);
-                _faceMeshCacheDirty = false;
-            }
-            catch (Exception e)
-            {
-                Debug.LogWarning(F("Warning.FaceMeshCacheSaveFailed", e.Message));
-            }
-        }
-
-        private static string GetFaceMeshCacheFilePath()
-        {
-            var projectRoot = Directory.GetParent(Application.dataPath)?.FullName ?? Application.dataPath;
-            return Path.Combine(projectRoot, "Library", "Aramaa", "OchibiChansConverterTool", FaceMeshCacheFileName);
-        }
-
-        private static bool TryParseHash128(string value, out Hash128 hash)
-        {
-            hash = default;
-            if (string.IsNullOrEmpty(value)) return false;
-
-            try
-            {
-                hash = Hash128.Parse(value);
-                return true;
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-        }
-
-        [Serializable]
-        private sealed class FaceMeshCacheFile
-        {
-            public List<FaceMeshCacheEntry> Entries = new List<FaceMeshCacheEntry>();
-        }
-
-        /// <summary>
-        /// フェイスメッシュキャッシュのシリアライズ用エントリです。
-        /// </summary>
-        [Serializable]
-        private sealed class FaceMeshCacheEntry
-        {
-            public string PrefabPath;
-            public string DependencyHash;
-            public string FaceMeshGuid;
-            public long FaceMeshLocalId;
-            public bool HasLocalId;
-            public bool HasFaceMesh;
-            public string PrefabGuid;
-            public string PrefabName;
-            public string FbxGuid;
-            public string FbxName;
-            public string FaceMeshAssetPath;
-        }
-
-        /// <summary>
-        /// メッシュ識別子（GUID と LocalId の組）です。
-        /// </summary>
-        private readonly struct MeshId
-        {
-            public MeshId(string guid, long localId, bool hasLocalId)
-            {
-                Guid = guid;
-                LocalId = localId;
-                HasLocalId = hasLocalId;
-            }
-
-            public string Guid { get; }
-            public long LocalId { get; }
-            public bool HasLocalId { get; }
-        }
-
-        /// <summary>
-        /// 顔メッシュの識別情報（GUID/LocalId + Prefab/FBX 識別子）です。
-        /// </summary>
-        private readonly struct FaceMeshSignature
-        {
-            public FaceMeshSignature(
-                MeshId meshId,
-                string prefabGuid,
-                string prefabName,
-                string fbxGuid,
-                string fbxName,
-                string faceMeshAssetPath)
-            {
-                MeshId = meshId;
-                PrefabGuid = prefabGuid;
-                PrefabName = prefabName;
-                FbxGuid = fbxGuid;
-                FbxName = fbxName;
-                FaceMeshAssetPath = faceMeshAssetPath;
-            }
-
-            public MeshId MeshId { get; }
-            public string PrefabGuid { get; }
-            public string PrefabName { get; }
-            public string FbxGuid { get; }
-            public string FbxName { get; }
-            public string FaceMeshAssetPath { get; }
-
-            public bool HasAnyIdentity =>
-                !string.IsNullOrEmpty(MeshId.Guid) ||
-                !string.IsNullOrEmpty(PrefabGuid) ||
-                !string.IsNullOrEmpty(PrefabName) ||
-                !string.IsNullOrEmpty(FbxGuid) ||
-                !string.IsNullOrEmpty(FbxName) ||
-                !string.IsNullOrEmpty(FaceMeshAssetPath);
-
-            public FaceMeshSignature WithPrefabInfo(string prefabGuid, string prefabName)
-            {
-                return new FaceMeshSignature(MeshId, prefabGuid, prefabName, FbxGuid, FbxName, FaceMeshAssetPath);
-            }
-        }
-
-        /// <summary>
-        /// Prefab の依存ハッシュと顔メッシュIDをまとめたキャッシュ情報です。
-        /// </summary>
-        private readonly struct CachedFaceMesh
-        {
-            public CachedFaceMesh(Hash128 dependencyHash, FaceMeshSignature faceMeshSignature, bool hasFaceMesh)
-            {
-                DependencyHash = dependencyHash;
-                FaceMeshSignature = faceMeshSignature;
-                HasFaceMesh = hasFaceMesh;
-            }
-
-            public Hash128 DependencyHash { get; }
-            public FaceMeshSignature FaceMeshSignature { get; }
-            public bool HasFaceMesh { get; }
         }
     }
 }
